[ct_telnet] Don't send extra newline after password
The telnet client used to send an extra newline after the
password. This caused an extra prompt to be sent back from the
server. Some test cases failed every now and then due to this - since
the second promt was confusing. This is now corrected, i.e. the client
does no longer send an extra newline after the password.