Add iso_week_to_date
Following http://en.wikipedia.org/wiki/ISO_week_date as reference add
iso_week_to_date/3 which takes ISO Week, year and day of the week and
finding the date associated with it. This is the inverse of
iso_week_number/1. This function returns the date representing the
iso week given to it