[ct_telnet] Fix test case 'server_speaks'
This test case simulates that the server spontaneously sends data. By
changing ct_telnet:send to ct_telnet_client:send_data this simulations
makes more sense - since client's buffers are not flushed.

Also removed the newline at the end of "echo_no_prompt" command. Since
newline is added automatically, the explicit newline at the end of the
command causes a prompt to be sent which is not expected.