Rewrite constraint handling
The internal representation for constraints (and object sets)
as produced by the parser was awkward, making further processing
convoluted. Here follows some examples of the old representation
for INTEGER constraints.

The constraint 1..2 is represented as:

  {'ValueRange',{1,2}}

If we extend the constraint like this:

  1..2, ...,

or like this:

  1..2, ..., 3

the representation would be:

  {{'ValueRange',{1,2}},[]}

and

  {{'ValueRange',{1,2}},{'SingleValue',3}}

respectively. Note that the pattern {A,B} will match all these
constraints.

When combining constraints using set operators:

  1..2 | 3..4 ^ 5..6

the representation will no longer be a tuple but a list:

  [{'ValueRange',{1..2}} union
   {'ValueRange',{3..4}} intersection
   {'ValueRange',{5..6}}]

The parse has full knowledge of the operator precedence; unfortunately,
the following pass (asn1ct_check) must also have the same knowledge
in order to correctly evaluate the constraints.

If we would change the order of the evaulation with round brackets:

  (1..2 | 3..4) ^ 5..6

there would be a nested listed in the representation:

  [[{'ValueRange',{1..2}} union {'ValueRange',{3..4}}]
   intersection {'ValueRange',{5..6}}]

We will change the representation to make it more explicit.
At the outer level, a constraint is always represented as

  {element_set,Root,Extension}

Extension will be 'none' if there is no extension, and 'empty' if
there is an empty extension. Root may also be 'empty' in an object set
if there are no objects in the root. Thus the constraints:

  1..2
  1..2, ...
  1..2, ..., 3

will be represented as:

  {element_set,{'ValueRange',{1,2}},none}
  {element_set,{'ValueRange',{1,2}},empty}
  {element_set,{'ValueRange',{1,2}},{'SingleValue',3}}

We will change the set operators too. This constraint:

  1..2 | 3..4 ^ 5..6

will be represented as:

  {element_set,
      {union,
        {'ValueRange',{1,2}},
        {intersection,
           {'ValueRange',{3,4}},
           {'ValueRange',{5,6}}},
       none}}

which is trivial to understand and evaluate. Similarly:

  (1..2 | 3..4) ^ 5..6

will be represented as:

  {element_set,
      {intersection,
        {union,{'ValueRange',{1,2}},{'ValueRange',{3,4}}},
        {'ValueRange',{5,6}}},
      none}