asn1ct_parser2: Simplify parse_Type/1
The parse_Type/1 calls various type parse functions. Most of those
functions return a #type record, but not all of them.  If a #type{}
record is not returned, parse_Type/1 will wrap the return value in a

We can simplify the code in parse_Type/1 if we make sure that the
type parsing functions called by parse_Type/1 always return
a #type{} record.