asn1ct_parser2: Eliminate all uses of old-style 'catch'
Most uses of 'catch' are not necessary. For example,
parse_or/2 is typically called like this:

    case catch parse_or(Tokens, Flist) of
       {'EXIT',Reason} ->
           exit(Reason);
       {asn1_error,_}=AsnErr ->
           throw(AsnErr);
       Result ->
           Result
    end.

Since {asn1_error,_} is always thrown (never returned) and
a successful is always returned (never thrown), the 'case' and
the 'catch' are not necessary. The code can be simplified too:

    parse_or(Tokens, Flist)

In a few cases, we will need to replace the 'catch' with
'try'...'catch'.