asn1ct_parser2: Remove expensive lookahead_assignment/1 function
In the parsing of a value assignment, such as:

   value INTEGER ::= 42

there is call to a function called lookahead_assignment/1 that
will ensure that the sequence of tokens that follows the value
is a valid assignment. The problem is that if the next assignment
is a value assignment, that too will look ahead to the next
assignment. That means that the complexity will be quadratic
if there are many value assignments following each other.

The reason for the test in the first place is unclear; my guess
is that it was an attempt to provide better error reporting.