Dialyzer to cache results of HiPE compilation
Cache the results of native compilation in
$XDG_CACHE_HOME/dialyzer_hipe_cache.  This can reduce the startup time
significantly.

If XDG_CACHE_HOME is unset, it defaults to $HOME/.cache.
See the XDG Base Directory spec for more details:
http://standards.freedesktop.org/basedir-spec/basedir-spec-0.7.html

Under that directory, create a subdirectory called
ARCH-VERSION-CHECKSUM, where ARCH is the HiPE architecture, VERSION is
the HiPE version number, and CHECKSUM is the return value of
hipe_bifs:system_crc/0.

Caching can be disabled by specifying --no_native_cache on the command
line.