[ct_cover] Fix paths of incl_dirs in cover spec
It seems like the commit 5a3c4668908254ee930c8db2e5cf9741945f9b2b also
broke the incl_dirs option and friends when given as relative paths.
When a cover spec contains a relative path it is looked-up in the
directory with the test results, not in the .cover file directory.