erts: Fix big-endian issue with printf_putc
On little-endian machines, doing &int will give the smallest
byte which is what we want to give to write. But on big-endian
it will give the highest byte, which will always be \000 here
which results in write never doing any writes.

So we have to cast c to an unsigned char before passing it to
write.