wx: On Mac keyboards use meta field as raw_ctrl
On mac and wxWidgets-3.0 the crtlDown field in keyboard and
mouse events are set when command button is pressed.

The ctrl key on mac keyboard is in a field called raw_control as this
would be backward incompatible to add we reuse the metaDown field as
indicator when the ctrl keyboard button is pressed.