Prevent zip:zip_open/1,2 from leaking ports
The case was discovered where a parent process would exit before closing
the zip file. The result was that a port would be left open
indefinitely, as the small zip server would not detect this condition.

By comparison, the file module will close the associated port when the
parent exits for any reason. This change would make the zip module more
consistent with the semantics of similar modules.

This change is breaking for any callers expecting to pass the handle to
another process for processing (assuming it exits).