core_lib: Handle patterns in map values
core_lib:is_var_used/2 would not consider a variable used in the
value of a map pattern such as:

  case Map of
    #{key := <<42:N>>} -> ok
  end

Here the variable 'N' would not be considered used.

It was assumed that there was no need to check map patterns at
all, since maps currently don't support variables in keys.