Move grouping of map constructions from v3_core to v3_kernel
When translating a function with map construction:

  f(A) ->
    B = b,
    C = c,
    #{A=>1,B=>2,C=>3}.

v3_core would break apart the map construction into three
parts because of the way the map instructions in BEAM work --
variable keys need to be in their own instruction.

In the example, constant propagation will turn two of the
keys to literal keys. But the initial breaking apart will
not be undone, so there will still be three map constructions:

  'f'/1 =
    fun (_cor0) ->
      let <_cor3> = ~{::<_cor0,1>}~
	in let <_cor4> = ~{::<'b',2>|_cor3}~
	  in ~{::<'c',3>|_cor4}~

It would be possible to complicate the sys_core_fold pass
to regroup map operations so that we would get:

  'f'/1 =
    fun (_cor0) ->
      let <_cor3> = ~{::<_cor0,1>}~
        in ~{::<'b',2>,::<'c',3>|_cor3}~

A simpler way that allows to simplify the translation is
to skip the grouping in v3_core and translate the function
to:

  'f'/1 =
    fun (_cor0) ->
      ~{::<_cor0,1>,::<'b',2>,::<'c',3>}~

We will then let v3_kernel do the grouping while translating
from Core Erlang to Kernel Erlang.