core_pp: Correct printing of map literals
A map key in a pattern would be incorrectly pretty-printed.
As an example, the pattern in:

  x() ->
    #{ #{ a => 3 } := 42 } = X.

would be pretty-printed as:

    <~{~<~{~<'a',3>}~,42>}~

instead of:

    <~{~<~{::<'a',3>}~,42>}~

When this problem has been corrected, the workaround for it in
cerl:ann_c_map/3 can be removed. The workaround was not harmless,
as it would cause the following map update to incorrectly succeed:

  (#{})#{a:=1}