Suppress warnings for expressions that are assigned to '_'
In c34ad2d5, the compiler learned to silence some warnings for
expressions that were explicitly assigned to the '_' variable,
as in this example:

  _ = list_to_integer(S),
  ok

That commit intentionally only made it possible to silence warnings
for BIFs that could cause an exception. Warnings would still be
produced for:

  _ = date(),
  ok

because date/0 can never fail and thus making the call completely
useless. The reasoning was that such warnings can always be
eliminated by eliminating the offending code.

While that is true, there is the question about rules and their
consistency. It is surprising that '_' can be used to silence
some warnings, but has no effect on other warnings.

Therefore, we will teach the compiler to silence warnings for
the following constructs:

  * Calls to safe BIFs such as date/0
  * Expressions that will cause an exception such as 'X/0'
  * Terms that are built but not used, such as '{x,X}'