v3_core: Use Core Erlang annotations in a type-safe way
Core Erlang annotations are supposed to be a list of terms.  v3_core
could temporarily stuff a record in the 'anno' field of a Core Erlang
record. That will cause Dialyzer warnings if we would tighten the
type specs for annotations. (We want to tighten the warnings in order
to catch more real problems.)

Avoid abusing the annotation by wrapping the entire Core Erlang
record in a #isimple{} record.

Reported-by: Kostis Sagonas