erts: Fix getifaddrs realloc ptr mismatch
When a buffer was exhausted and subsequently a realloc, we could get
an invalid pointer.

For this to occur we would need to have a realloc to lower adresses.

The symptom would be garbage returned from erlang:port_control(Port, 25, [])
(prim_inet:getifaddrs(Port) resulting in a badarg) or a segmentation fault.