v3_core: Suppress compiler-generated calls in guards
Compiling the following function:

  f(V) when not (bar and V) -> true;   %Line 4
  f(_) -> false.

would produce the following warnings:

  no_file: Warning: the call to is_boolean/1 has no effect
  t.erl:4: Warning: the guard for this clause evaluates to 'false'
  t.erl:4: Warning: use of operator '=:=' has no effect

Two of the warnings refer to calls to is_boolean/1 and '=:='/2 which
v3_core added when translating the code to Core Erlang. The only
relevant warning is:

  t.erl:4: Warning: the guard for this clause evaluates to 'false'

Suppress the other two warning by marking the compiler-generated
calls with a 'compiler_generated' annotation.