trace_bif_SUITE: Ensure that a call to time/0 is not removed
Assigning the result of a BIF call to a variable used to be
sufficient to prevent the call from being optimized away. With
the more aggressive optimization that will be introduced in
a future commit it will not be sufficient.

Matching the return value will prevent the compiler from doing
the optimization.