Correct unsafe optimization of '==' and '/='
Since '=:=' is cheaper than '==', the compiler tries to replace
'==' with '=:=' if the result of comparison will be the same.

As an example:

  V == {a,b}

can be rewritten to:

  V =:= {a,b}

since the literal on the right side contains no numeric values
that '==' would compare differently to '=:='.

With the introduction of maps, we will need to take them into
account. Since the comparison of maps is planned to change in 18.0,
we will be very conservative and only do the optimization if
both keys and values are non-numeric.