remove perfctr support
Perfctr is a Linux kernel extension that allows programmatic access
to the performance monitoring counters found in most current CPUs.
However, development of perfctr ceased after 2010, and it cannot be
used with Linux kernels newer than 2.6.32.

Therefore the perfctr support code in the Erlang VM is effectively
dead code, so this patch removes it.