[cover] Improve performance
Add functions for cover compilation and analysis on multiple
files. This allows for more parallelisation.

All functions for cover compilation can now take a list of
modules/files.

cover:analyse/analyze and cover:analyse_to_file/analyze_to_file can be
called without the Modules arguement in order to analyse all cover
compiled and imported modules, or with a list of modules.

Also, the number of lookups in ets tables is reduced, which has also
improved the performance when analysing and resetting cover data.